sqs-large-message/sqs-large-message/pyproject.toml
[project]
name = "sqs-large-message"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amazon-sqs-extended-client>=1.0.1",
    "boto3>=1.40.15",
]
